Fixes broken custom heap size execution and configures dependent task
  * Adds the dependsOn parameter to specify the task on which the
javarepl should depend.  It defaults to the testClasses task.
  * Fixes creation of the command line when no heap size and/or stack
size is specified.  Previously, blank values were put in the list passed
on to the ProcessBuilder which caused JavaREPL to fail.  Now, nothing is
added to list when these parameters are not specified.
  * Moves the execution of the logic of the task to the doLast method to
start JavaREPL after dependent tasks have successfully executed
  * Adds the JavaReplPluginExtension.validate method to verify the
values of the plugin parameters
  * Adds test cases for invalid version, configurations, and dependsOn
values, as well as, the behavior of the new dependsOn parameter